# Several performance improvements