Before

```
BasicBench_AdaURL     175465 ns       175464 ns         3976 GHz=3.26934 cycle/byte=24.0253 cycles/url=746.974 instructions/byte=100.423 instructions/cycle=4.17988 instructions/ns=13.6654 instructions/url=3.12226k ns/url=228.478 speed=134.136M/s time/byte=7.45512ns time/url=231.788ns url/s=4.31428M/s
```

After

```
BasicBench_AdaURL     172468 ns       172467 ns         4089 GHz=3.24986 cycle/byte=23.5887 cycles/url=733.4 instructions/byte=99.9744 instructions/cycle=4.23823 instructions/ns=13.7737 instructions/url=3.10832k ns/url=225.671 speed=136.466M/s time/byte=7.32781ns time/url=227.83ns url/s=4.38924M/s
```