@anonrig 2023-02-13
I only see micro changes in the `wpt_bench` not on `bench`. Might be not worth the changes. @lemire. Feel free to close if you seem fit