@lemire 2023-02-13
@anonrig I skeptical regarding one change... where you drop the `rfind` and replaced it by your own loop. It might be that it is faster... but this code...

```
if (!path.empty()) {
      path.erase(path.rfind('/'));
}
```

should be decent... It will trigger a function call (to rfind). I am not requesting that you change anything, but how sure are you that what you did resulted in better performance?

If the difference is too tiny to measure, I would simply revert to the simpler original code.

If you are convinced, merge away !!!